PR3 review-cycle fixes
Frontend wiring + cleanup from internal review of the heuristic policy
enumeration API.

- Plumb pluginName/pluginNameLen through the IHipdnnBackend interface,
  BackendWrapper, IncompatibleBackend stub, MockHipdnnBackend, and the
  HeuristicPolicyInfo two-call query so the frontend matches the merged
  PR2 backend C ABI (added pluginName field to HeuristicPolicyInfo).
- Restore (override) on 5 MOCK_METHODs that lost it, and add it to the
  two new heuristic-policy mocks so non-virtual overrides fail to build.
- Report backend errors via getErrorString(status) instead of the raw
  integer; harden the buffer-to-string conversion to use the reported
  length and strip a trailing NUL only when present.
- Verify policy IDs against policyNameToId (from PolicyNames.hpp),
  rename the integration suite to IntegrationGpuHeuristicPolicyEnumeration
  per the GPU-fixture naming rule, assert pluginName is non-empty,
  include it in the logging test output, and drop the flaky 100ms
  performance assertion.
- Tighten serializeDeviceProperties doc to say the returned vector owns
  its bytes and the plugin-const-data wrapper aliases it.
- Fix SPDX double-space and route reinterpret_cast sentinel pointers
  through uintptr_t in TestEngineHeuristicDescriptorAdditional.cpp.

+++ projects/hipdnn/backend/tests/descriptors/TestEngineHeuristicDescriptorAdditional.cpp
@@ -1,5 +1,5 @@
 // Copyright © Advanced Micro Devices, Inc., or its affiliates.
-// SPDX-License-Identifier: MIT
+// SPDX-License-Identifier:  MIT
 
 /**
  * @file TestEngineHeuristicDescriptorAdditional.cpp
@@ -83,8 +83,10 @@ public:
         const int64_t staticOrderingPolicyId
             = hipdnn_data_sdk::utilities::engineNameToId("SelectionHeuristic::StaticOrdering");
 
-        auto mockHandle = reinterpret_cast<hipdnnHeuristicHandle_t>(0x1234);
-        auto mockDescriptor = reinterpret_cast<hipdnnHeuristicPolicyDescriptor_t>(0x5678);
+        // Sentinel pointer values — the mocks only compare them by identity, never deref.
+        auto mockHandle = reinterpret_cast<hipdnnHeuristicHandle_t>(static_cast<uintptr_t>(0x1234));
+        auto mockDescriptor
+            = reinterpret_cast<hipdnnHeuristicPolicyDescriptor_t>(static_cast<uintptr_t>(0x5678));
 
         // For custom/unknown policy IDs, return nullptr (set up general expectations first)
         EXPECT_CALL(*_mockHeuristicPluginResourceManager, getPluginForPolicyId(_))
